♻️(frontend) AppShell props navigation and use current user store
