+++ src/web/presentation/frontend/src/app/App.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import { RouterView } from 'vue-router'
+import { APP_NAVIGATION } from '@/app/navigation'
 import CspToaster from '@/components/base/CspToast/CspToaster.vue'
 import ErrorBoundary from '@/components/ErrorBoundary.vue'
 import AppShell from '@/components/layout/AppShell.vue'
@@ -8,7 +9,7 @@ import AppShell from '@/components/layout/AppShell.vue'
 <template>
   <div class="ats-app">
     <CspToaster>
-      <AppShell>
+      <AppShell :navigation="APP_NAVIGATION">
         <ErrorBoundary>
           <RouterView />
         </ErrorBoundary>
