+++ src/web/presentation/frontend/src/app/navigation.ts
@@ -1,16 +1,6 @@
-export interface NavItem {
-  icon: string
-  label: string
-  /** Nom de route vue-router cible. */
-  to: string
-}
+import type { NavGroup } from '@/components/layout/AppShell.types'
 
-export interface NavGroup {
-  label: string
-  items: NavItem[]
-}
-
-export const ATS_NAVIGATION: NavGroup[] = [
+export const APP_NAVIGATION: NavGroup[] = [
   {
     label: 'Pilotage',
     items: [
