+++ src/web/presentation/frontend/src/components/layout/AppShell.types.ts
@@ -1,0 +1,11 @@
+export interface NavItem {
+  icon: string
+  label: string
+  /** Nom de route vue-router cible. */
+  to: string
+}
+
+export interface NavGroup {
+  label: string
+  items: NavItem[]
+}
